requirements.txt
spacy==3.8.2
llama-cpp-python>=0.3.2
transformers==4.46.3
torch==2.5.1
numpy>=1.26.2
scikit-learn>=1.3.2
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=1.4.0
nlpaug>=1.1.3
tqdm>=4.67.0
# python -m spacy download en_core_web_sm